requirements.txt
opencv-python>=4.8.0
cmake
dlib-bin
numpy>=1.24.0
Flask>=2.3.0
Flask-SocketIO>=5.3.0
pystray>=0.19.0
Pillow>=10.0.0
scipy>=1.10.0
